requirements.txt
Django==5.2.5
djangorestframework==3.15.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.2.0
psycopg[binary]==3.2.1
